+++ Snowplow/Events/SPEcommerce.m
@@ -1,0 +1,110 @@
+//
+//  SPEcommerce.m
+//  Snowplow
+//
+//  Created by Alex Benini on 14/02/2020.
+//  Copyright © 2020 Snowplow Analytics. All rights reserved.
+//
+
+#import "SPEcommerce.h"
+
+#import "Snowplow.h"
+#import "SPUtilities.h"
+#import "SPPayload.h"
+
+@implementation SPEcommerce {
+    NSString * _orderId;
+    NSNumber * _totalValue;
+    NSString * _affiliation;
+    NSNumber * _taxValue;
+    NSNumber * _shipping;
+    NSString * _city;
+    NSString * _state;
+    NSString * _country;
+    NSString * _currency;
+    NSArray *  _items;
+}
+
++ (instancetype) build:(void(^)(id<SPEcommTransactionBuilder>builder))buildBlock {
+    SPEcommerce* event = [SPEcommerce new];
+    if (buildBlock) { buildBlock(event); }
+    [event preconditions];
+    return event;
+}
+
+- (id) init {
+    self = [super init];
+    return self;
+}
+
+- (void) preconditions {
+    [SPUtilities checkArgument:([_orderId length] != 0) withMessage:@"OrderId cannot be nil or empty."];
+    [SPUtilities checkArgument:(_items != nil) withMessage:@"Items cannot be nil."];
+    [SPUtilities checkArgument:(_totalValue != nil) withMessage:@"TotalValue cannot be nil."];
+    [self basePreconditions];
+}
+
+// --- Builder Methods
+
+- (void) setOrderId:(NSString *)orderId {
+    _orderId = orderId;
+}
+
+- (void) setTotalValue:(double)totalValue {
+    _totalValue = [NSNumber numberWithDouble:totalValue];
+}
+
+- (void) setAffiliation:(NSString *)affiliation {
+    _affiliation = affiliation;
+}
+
+- (void) setTaxValue:(double)taxValue {
+    _taxValue =  [NSNumber numberWithDouble:taxValue];
+}
+
+- (void) setShipping:(double)shipping {
+    _shipping =  [NSNumber numberWithDouble:shipping];
+}
+
+- (void) setCity:(NSString *)city {
+    _city = city;
+}
+
+- (void) setState:(NSString *)state {
+    _state = state;
+}
+
+- (void) setCountry:(NSString *)country {
+    _country = country;
+}
+
+- (void) setCurrency:(NSString *)currency {
+    _currency = currency;
+}
+
+- (void) setItems:(NSArray *)items {
+    _items = items;
+}
+
+// --- Public Methods
+
+- (SPPayload *) getPayload {
+    SPPayload *pb = [[SPPayload alloc] init];
+    [pb addValueToPayload:kSPEventEcomm forKey:kSPEvent];
+    [pb addValueToPayload:_orderId forKey:kSPEcommId];
+    [pb addValueToPayload:[NSString stringWithFormat:@"%.02f", [_totalValue doubleValue]] forKey:kSPEcommTotal];
+    [pb addValueToPayload:_affiliation forKey:kSPEcommAffiliation];
+    [pb addValueToPayload:[NSString stringWithFormat:@"%.02f", [_taxValue doubleValue]] forKey:kSPEcommTax];
+    [pb addValueToPayload:[NSString stringWithFormat:@"%.02f", [_shipping doubleValue]] forKey:kSPEcommShipping];
+    [pb addValueToPayload:_city forKey:kSPEcommCity];
+    [pb addValueToPayload:_state forKey:kSPEcommState];
+    [pb addValueToPayload:_country forKey:kSPEcommCountry];
+    [pb addValueToPayload:_currency forKey:kSPEcommCurrency];
+    return [self addDefaultParamsToPayload:pb];
+}
+
+- (NSArray *) getItems {
+    return _items;
+}
+
+@end
