+++ Snowplow/Events/SPEcommerceItem.m
@@ -1,0 +1,90 @@
+//
+//  SPEcommerceItem.m
+//  Snowplow
+//
+//  Created by Alex Benini on 14/02/2020.
+//  Copyright © 2020 Snowplow Analytics. All rights reserved.
+//
+
+#import "SPEcommerceItem.h"
+
+#import "Snowplow.h"
+#import "SPUtilities.h"
+#import "SPPayload.h"
+
+@implementation SPEcommerceItem {
+    NSString * _itemId;
+    NSString * _sku;
+    NSNumber * _price;
+    NSNumber * _quantity;
+    NSString * _name;
+    NSString * _category;
+    NSString * _currency;
+}
+
++ (instancetype) build:(void(^)(id<SPEcommTransactionItemBuilder>builder))buildBlock {
+    SPEcommerceItem* event = [SPEcommerceItem new];
+    if (buildBlock) { buildBlock(event); }
+    [event preconditions];
+    return event;
+}
+
+- (id) init {
+    self = [super init];
+    return self;
+}
+
+- (void) preconditions {
+    [SPUtilities checkArgument:([_itemId length] != 0) withMessage:@"ItemId cannot be nil or empty."];
+    [SPUtilities checkArgument:([_sku length] != 0) withMessage:@"SKU cannot be nil or empty."];
+    [SPUtilities checkArgument:(_price != nil) withMessage:@"Price cannot be nil."];
+    [SPUtilities checkArgument:(_quantity != nil) withMessage:@"Quantity cannot be nil."];
+    [self basePreconditions];
+}
+
+// --- Builder Methods
+
+- (void) setItemId:(NSString *)itemId {
+    _itemId = itemId;
+}
+
+- (void) setSku:(NSString *)sku {
+    _sku = sku;
+}
+
+- (void) setPrice:(double)price {
+    _price = [NSNumber numberWithDouble:price];
+}
+
+- (void) setQuantity:(NSInteger)quantity {
+    _quantity = [NSNumber numberWithLong:quantity];
+}
+
+- (void) setName:(NSString *)name {
+    _name = name;
+}
+
+- (void) setCategory:(NSString *)category {
+    _category = category;
+}
+
+- (void) setCurrency:(NSString *)currency {
+    _currency = currency;
+}
+
+// --- Public Methods
+
+- (SPPayload *) getPayload {
+    SPPayload *pb = [[SPPayload alloc] init];
+    [pb addValueToPayload:kSPEventEcommItem forKey:kSPEvent];
+    [pb addValueToPayload:_itemId forKey:kSPEcommItemId];
+    [pb addValueToPayload:_sku forKey:kSPEcommItemSku];
+    [pb addValueToPayload:_name forKey:kSPEcommItemName];
+    [pb addValueToPayload:_category forKey:kSPEcommItemCategory];
+    [pb addValueToPayload:[NSString stringWithFormat:@"%.02f", [_price doubleValue]] forKey:kSPEcommItemPrice];
+    [pb addValueToPayload:[NSString stringWithFormat:@"%ld", [_quantity longValue]] forKey:kSPEcommItemQuantity];
+    [pb addValueToPayload:_currency forKey:kSPEcommItemCurrency];
+    return [self addDefaultParamsToPayload:pb];
+}
+
+@end
