+++ Snowplow/Events/SPEvents.h
@@ -1,0 +1,29 @@
+//
+//  SPEvents.h
+//  Snowplow
+//
+//  Created by Alex Benini on 14/02/2020.
+//  Copyright © 2020 Snowplow Analytics. All rights reserved.
+//
+
+#ifndef Header_h
+#define Header_h
+
+#import "SPEvent.h"
+#import "SPPageView.h"
+#import "SPStructured.h"
+#import "SPUnstructured.h"
+#import "SPScreenView.h"
+#import "SPConsentWithdrawn.h"
+#import "SPConsentDocument.h"
+#import "SPConsentGranted.h"
+#import "SPTiming.h"
+#import "SPEcommerce.h"
+#import "SPEcommerceItem.h"
+#import "SPNotificationContent.h"
+#import "SPPushNotification.h"
+#import "SPForeground.h"
+#import "SPBackground.h"
+#import "SNOWError.h"
+
+#endif /* Header_h */
