+++ Snowplow/Events/SPStructured.m
@@ -1,0 +1,76 @@
+//
+//  SPStructured.m
+//  Snowplow
+//
+//  Created by Alex Benini on 14/02/2020.
+//  Copyright © 2020 Snowplow Analytics. All rights reserved.
+//
+
+#import "SPStructured.h"
+
+#import "Snowplow.h"
+#import "SPUtilities.h"
+#import "SPPayload.h"
+
+@implementation SPStructured {
+    NSString * _category;
+    NSString * _action;
+    NSString * _label;
+    NSString * _property;
+    NSNumber * _value;
+}
+
++ (instancetype) build:(void(^)(id<SPStructuredBuilder>builder))buildBlock {
+    SPStructured* event = [SPStructured new];
+    if (buildBlock) { buildBlock(event); }
+    [event preconditions];
+    return event;
+}
+
+- (id) init {
+    self = [super init];
+    return self;
+}
+
+- (void) preconditions {
+    [SPUtilities checkArgument:([_category length] != 0) withMessage:@"Category cannot be nil or empty."];
+    [SPUtilities checkArgument:([_action length] != 0) withMessage:@"Action cannot be nil or empty."];
+    [self basePreconditions];
+}
+
+// --- Builder Methods
+
+- (void) setCategory:(NSString *)category {
+    _category = category;
+}
+
+- (void) setAction:(NSString *)action {
+    _action = action;
+}
+
+- (void) setLabel:(NSString *)label {
+    _label = label;
+}
+
+- (void) setProperty:(NSString *)property {
+    _property = property;
+}
+
+- (void) setValue:(double)value {
+    _value = [NSNumber numberWithDouble:value];
+}
+
+// --- Public Methods
+
+- (SPPayload *) getPayload {
+    SPPayload *pb = [[SPPayload alloc] init];
+    [pb addValueToPayload:kSPEventStructured forKey:kSPEvent];
+    [pb addValueToPayload:_category forKey:kSPStuctCategory];
+    [pb addValueToPayload:_action forKey:kSPStuctAction];
+    [pb addValueToPayload:_label forKey:kSPStuctLabel];
+    [pb addValueToPayload:_property forKey:kSPStuctProperty];
+    [pb addValueToPayload:[NSString stringWithFormat:@"%.17g", [_value doubleValue]] forKey:kSPStuctValue];
+    return [self addDefaultParamsToPayload:pb];
+}
+
+@end
