+++ Snowplow/Events/SPTiming.m
@@ -1,0 +1,74 @@
+//
+//  SPTiming.m
+//  Snowplow
+//
+//  Created by Alex Benini on 14/02/2020.
+//  Copyright © 2020 Snowplow Analytics. All rights reserved.
+//
+
+#import "SPTiming.h"
+
+#import "Snowplow.h"
+#import "SPUtilities.h"
+#import "SPSelfDescribingJson.h"
+
+@implementation SPTiming {
+    NSString * _category;
+    NSString * _variable;
+    NSNumber * _timing;
+    NSString * _label;
+}
+
++ (instancetype) build:(void(^)(id<SPTimingBuilder>builder))buildBlock {
+    SPTiming* event = [SPTiming new];
+    if (buildBlock) { buildBlock(event); }
+    [event preconditions];
+    return event;
+}
+
+- (id) init {
+    self = [super init];
+    return self;
+}
+
+- (void) preconditions {
+    [SPUtilities checkArgument:([_category length] != 0) withMessage:@"Category cannot be nil or empty."];
+    [SPUtilities checkArgument:([_variable length] != 0) withMessage:@"Variable cannot be nil or empty."];
+    [SPUtilities checkArgument:(_timing != nil) withMessage:@"Timing cannot be nil."];
+    [self basePreconditions];
+}
+
+// --- Builder Methods
+
+- (void) setCategory:(NSString *)category {
+    _category = category;
+}
+
+- (void) setVariable:(NSString *)variable {
+    _variable = variable;
+}
+
+- (void) setTiming:(NSInteger)timing {
+    _timing = [NSNumber numberWithLong:timing];
+}
+
+- (void) setLabel:(NSString *)label {
+    _label = label;
+}
+
+// --- Public Methods
+
+- (SPSelfDescribingJson *) getPayload {
+    NSMutableDictionary * event = [[NSMutableDictionary alloc] init];
+    [event setObject:_category forKey:kSPUtCategory];
+    [event setObject:_variable forKey:kSPUtVariable];
+    [event setObject:_timing forKey:kSPUtTiming];
+    if (_label != nil) {
+        [event setObject:_label forKey:kSPUtLabel];
+    }
+
+    return [[SPSelfDescribingJson alloc] initWithSchema:kSPUserTimingsSchema
+                                                andData:event];
+}
+
+@end
