+++ Snowplow/Events/SPTrackerEvent.m
@@ -1,0 +1,18 @@
+//
+//  SPTrackerEvent.m
+//  Snowplow
+//
+//  Created by Alex Benini on 13/02/2020.
+//  Copyright © 2020 Snowplow Analytics. All rights reserved.
+//
+
+#import "SPTrackerEvent.h"
+
+@implementation SPTrackerEvent
+
+- (instancetype)initWithEvent:(SPEvent *)event {
+//    self.payload = event get
+    return nil;
+}
+
+@end
