+++ Snowplow/SPSession.m
@@ -25,7 +25,7 @@
 #import "SPUtilities.h"
 #import "SPWeakTimerTarget.h"
 #import "SPTracker.h"
-#import "SPEvent.h"
+#import "SPEvents.h"
 
 #if SNOWPLOW_TARGET_IOS
 #import <UIKit/UIKit.h>
